refactor: [#1541] create folder for mod
More submods will be included inside.
