+++ packages/tracker-core/tests/common/test_env.rs
@@ -5,7 +5,7 @@ use aquatic_udp_protocol::AnnounceEvent;
 use bittorrent_primitives::info_hash::InfoHash;
 use bittorrent_tracker_core::announce_handler::PeersWanted;
 use bittorrent_tracker_core::container::TrackerCoreContainer;
-use bittorrent_tracker_core::statistics::persisted_metrics::load_persisted_metrics;
+use bittorrent_tracker_core::statistics::persisted::load_persisted_metrics;
 use tokio::task::yield_now;
 use torrust_tracker_configuration::Core;
 use torrust_tracker_metrics::label::LabelSet;
